🤡  Refactoring export method
