+++ example/package.json
@@ -14,7 +14,7 @@
     "@types/react": "^18.0.27",
     "@types/react-dom": "^18.0.10",
     "axios": "^1.3.2",
-    "flourish-web-sdk-react": "^1.0.7",
+    "flourish-web-sdk-react": "^1.0.8",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
