+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "flourish-web-sdk-react",
-  "version": "1.0.7",
+  "version": "1.0.8",
   "description": "Flourish React Web SDK",
   "main": "./dist/cjs/index.js",
   "module": "./dist/esm/index.js",
