--- src/components/index.ts
@@ -1,2 +1,0 @@
-export { default as Flourish } from "./Flourish";
-export { default as initializer } from "./Initializer";
